Rename positions2 to positions
